Refactoring: Allow GreenfieldExtensions to be used by plugins
