+++ BTCPayServer/Hosting/Startup.cs
@@ -3,6 +3,7 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Net;
+using BTCPayServer.Abstractions.Extensions;
 using BTCPayServer.Configuration;
 using BTCPayServer.Controllers.Greenfield;
 using BTCPayServer.Data;
